reto_data_engineer/requirements.txt
pandas
python-dateutil
sqlalchemy
psycopg2-binary
pyyaml
